% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_algorithm.R
\name{des_algorithm}
\alias{des_algorithm}
\title{DES optimization algorithm minimizing specified function}
\usage{
des_algorithm(par, fn, lower, upper, max_eval, tunable_parameters = NULL)
}
\arguments{
\item{par}{Initial parameter combination .}

\item{fn}{Function to minimize.}

\item{lower}{Lower bounds of search space.}

\item{upper}{Upper bounds of search space.}

\item{max_eval}{Maximium number of function evaluations to perform.}

\item{tunable_parameters}{list of tunable parameters which overrides hardcoded ones. Defaults to NULL (no overriding)}
}
\value{
List consisting of keys par - coordinates of best solution found and value - value of best solution .
}
\description{
DES optimization algorithm minimizing specified function
}
