% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{ds}
\alias{ds}
\title{create box plot, histogram, t-test and numeric summary}
\usage{
ds(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
create box plot, histogram, t-test and numeric summary
}
