% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpel.R
\name{simpel}
\alias{simpel}
\title{A Simpel Function}
\usage{
simpel(ak, lm, rd, rsc, rn, ongoing = TRUE)
}
\arguments{
\item{ak}{Entity arrival rate}

\item{lm}{service duration rate}

\item{rd}{rounding value}

\item{rsc}{amount of resources}

\item{rn}{number of simulation runs}

\item{ongoing}{record the arrival data of entities that have not yet left the system. Default to TRUE.}
}
\description{
This function allows you to simulate simple queueing system with single server (single queue)
}
\examples{
simpel()
}
\keyword{single_queue}
