% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpel3.R
\name{simpel3}
\alias{simpel3}
\title{A Simpel3 Function}
\usage{
simpel3(tri, unif, rn, res = TRUE)
}
\arguments{
\item{tri}{triangular distribution parameters}

\item{unif}{uniform distribution parameters}

\item{rn}{number of simulation run}

\item{res}{whether you want to display data resources. Default to TRUE}
}
\description{
This function allows you to simulate queuing system with more than one serial server, and each server has its own queue
}
\examples{
simpel3()
}
\keyword{every_server_has_each_queue}
