% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpel2.R
\name{simple2}
\alias{simple2}
\title{A Simpel2 Function}
\usage{
simple2(rate, awal, times, attributes = TRUE)
}
\arguments{
\item{rate}{entity arrival rate}

\item{awal}{earliest arrival time}

\item{times}{repetition}

\item{attributes}{whether your want to display data attributes. Default to TRUE}
}
\description{
This function allows you to simulate Simple queuing system with more than one single entity type (single queue)
}
\examples{
simpel2()
}
\keyword{more_than_one_single_entity}
